requirements/development.txt
absl-py==1.0.0
alabaster==0.7.12
asgiref==3.4.1
astunparse==1.6.3
Babel==2.9.1
backports.entry-points-selectable==1.1.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.8
coverage==6.2
cycler==0.11.0
distlib==0.3.3
Django==3.2.9
docutils==0.17.1
filelock==3.4.0
flake8==4.0.1
flatbuffers==2.0
fonttools==4.28.3
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h5py==3.6.0
idna==3.3
imagesize==1.3.0
importlib-metadata==4.10.0
imutils==0.5.4
isort==5.10.1
Jinja2==3.0.3
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
mccabe==0.6.1
numpy==1.21.4
oauthlib==3.1.1
opencv-python==4.5.4.60
opt-einsum==3.3.0
packaging==21.3
Pillow==8.4.0
platformdirs==2.4.0
pluggy==1.0.0
protobuf==3.19.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.10.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
snowballstemmer==2.2.0
sqlparse==0.4.2
tensorflow-macos
termcolor==1.1.0
toml==0.10.2
tox==3.24.4
typing_extensions==4.0.1
urllib3==1.26.7
virtualenv==20.10.0
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.6.0
